The following modifications require the viewr to pass ?lastpost=<post_name> in the url, and all posts from the post with that post name up to the first post will be shown.
If a post with post_name lastpost does not exist, no posts are shown.
This is designed to be an archive for webcomics, designed so that only people who know the link can read it up to that point.

If you change $renderall = false; to $renderall = true; in the line below, the code will revert to the default and show everything.